+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Allow to remove Live Entries when deleting assets or directories referenced by Live Entries.
+
 ## [1.0.1] - 2023-06-03
 
 ### Changed